Inserted case tests

+++ src/GLWidget.cpp
@@ -89,11 +89,8 @@ void GLWidget::initTextures()
 
 void GLWidget::resizeGL(int w, int h)
 {
-    // Calculate aspect ratio
-    qreal aspect = qreal(w) / qreal(h ? h : 1);
-
-    // Set near plane to 3.0, far plane to 7.0, field of view 45 degrees
-    const qreal zNear = 0.1, zFar = 10240.0, fov = 60.0;
+	float aspect = float(w) / float(h ? h : 1);
+	const float zNear = 0.1f, zFar = 10240.0f, fov = 60.0f;
 
     // Reset projection
     projection.setToIdentity();
